Rearranges everything, yes, again
