+++ src/components/Identify/ByImage/index.js
