+++ src/components/Identify/ByImage/index.test.js
@@ -1,15 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Scan from './index.js';
-import AppWrapper from 'AppWrapper';
+import ScanImage from './ScanImage';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
-  ReactDOM.render(
-    <AppWrapper>
-      <Scan />
-    </AppWrapper>,
-    div
-  );
+  ReactDOM.render(<ScanImage />, div);
   ReactDOM.unmountComponentAtNode(div);
 });
